+++ app/services/invoices/update_service.rb
@@ -41,6 +41,6 @@ class Invoices::UpdateService < BaseService
     return unless invoice.invoice_type == 'credit'
     return unless status == 'succeeded'
 
-    PrepaidCreditJob.perform_later(invoice)
+    Invoices::PrepaidCreditJob.perform_later(invoice)
   end
 end
